Final gui input checker plus reader and updater.
The gui at this stage is perfect.
It upload all the variables from the excel file.
It represent all the variable to the gui with all options.
It update variable dictionary plus textboxes according to input changes
(both textboxes and comboboxes).
It also makes the parameter vector or static according.
Also checks for the inputs and updated among corrected inputs.