Added retrial if the trial is not in the decision stage - randomally - checked.
Added in the alogorith of taking the next step the fact that if the
trial was not going to the decision/response stage than the experiment
should retrial it in randomly time in the roound of trials.
Also updated the counter in the globl gui variable for counting the
remaining trials followed by that factor.