Added a seperate function for updaintg jbi file and start moving the robot
for more precisely simulatenously working of the leds and the robot -
now it is o.k - it's about 50ms difference in the begining and the end.

TODO:
Take care if need to put the update JBI file function call in other
stage in another thread instead of in the duration stage because it
takes about 500ms and maybe we can over it in other stage.